PKG: Add build-info and move to `src` structure
